gitconfig: replace fa and fas aliases with sync
The new sync alias attempts to perfectly mirror the remote withing our
repository. All tags and branches will be fetched and deleted remote
branches will be pruned.

Note that this does not recurse by default. Because this is a wrapper
around fetch, --recurse-submodules can be given to achieve that
behavior.