remove files that are not needed for msysgit
MSYS is a minimal POSIX-like enviornment that is not fully POSIX
compliant. MSYS provides the core utilites git needs to function and not
much else. Therefore, most of these configuration files are unnecessary.

An argument could be made to keep vimrc since MSYS includes a
feature-reduced build of vim. However, I have decided that since vim is
being used only as a commit editor and not a general text editor, it is
not worth the hassle of determining if the vimrc file is fully portable
to MSYS.