remove gitignore
A global ignore list is dangerous because it makes assumptions about all
policies. This has burned me on several occasions.

For example, the global ignore list included static libraries as these
are normally build artifacts. However, if you are tracking an SDK in
git, then the static libraries are important and need to be tracked in
the repository.