land a bunch of changes to configure awesome
Xresources now uses the preprocessor for consistent colors across
settings. Xft is used so that fonts are scaled nicely on various
resolutions.

xinitrc gets installed as both ~/.xinitrc and ~/.xsession. This is
because startx uses the former while display managers use the latter.

awesome's rc.lua is still a work in progress, but is has reached a
useable state (i.e. I've been using it at work for a couple of weeks
now). There are a couple notes and commented out things that need to be
fixed. The menu that is launched with Mod4+Escape is incomplete.