awesome: use generic classes for Gecko browsers
All Gecko browsers (Firefox, Iceweasel and Seamonkey) use the
"Navigator" window class. All Gecko mail clients (Icedove, Thunderbird
and Seamonkey) use the "Mail" window class.

The primary motivation for this change is to force different Seamonkey
windows to different tags. This is not possible if the "Seamonkey" class
is used as this is used by all of the program's windows.